+++ clang-tools-extra/test/clang-tidy/checkers/daedalean-use-noexcept.cpp
@@ -1,0 +1,28 @@
+// RUN: %check_clang_tidy %s daedalean-use-noexcept %t
+
+class S {
+public:
+  S();
+  // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: Method 'S' should be noexcept [daedalean-use-noexcept]
+  S(const S &)
+  noexcept = default;
+
+  bool operator==(const S &) const noexcept;
+  bool operator!=(const S &) const;
+  // CHECK-MESSAGES: :[[@LINE-1]]:8: warning: Method 'operator!=' should be noexcept [daedalean-use-noexcept]
+};
+
+void f() noexcept;
+
+void f2();
+// CHECK-MESSAGES: :[[@LINE-1]]:6: warning: Function 'f2' should be noexcept [daedalean-use-noexcept]
+//
+void func() noexcept {
+  auto noexceptLambda = []() noexcept {
+
+  };
+
+  auto lambda = []() {
+    // CHECK-MESSAGES: :[[@LINE-1]]:17: warning: Lambda should be noexcept [daedalean-use-noexcept]
+  };
+}